rename maybe_from_str to maybe_quoted
