+++ proto/src/serializers.rs
@@ -3,7 +3,7 @@
 pub mod bytes;
 pub mod cow_str;
 pub mod from_str;
-pub mod maybe_from_str;
+pub mod maybe_quoted;
 pub mod null_default;
 pub mod option_protobuf_duration;
 pub mod option_timestamp;
